move javafx module to root
